Consolidate test fixtures and migrate holidays stops to separate spec
