+++ app/client/fixtures/payment.ts
@@ -1,0 +1,20 @@
+import { StripeSetupIntent } from '../../shared/stripeSetupIntent';
+
+export const stripeSetupIntent: StripeSetupIntent = {
+	id: 'seti_0KLFtUItVxyc3M6nBXnYb2jO',
+	client_secret:
+		'seti_0KLFtUItVxyc3Q6nBXnYb2jO_secret_L1IUioSNMNThetlMQnVtbCJu0Gj2cq1M',
+};
+
+export const executePaymentUpdateResponse = {
+	type: 'visa',
+	last4: '4242',
+	expiryMonth: 4,
+	expiryYear: 2024,
+};
+
+export const ddPaymentMethod = {
+	accountName: 'JON R HEE',
+	accountNumber: '****9911',
+	sortCode: '200000',
+};
