+++ app/client/fixtures/productDetail.ts
@@ -264,3 +264,138 @@ export const newspaperVoucherPaypal: ProductDetail = {
 	isTestUser: false,
 	key: '1639394814906',
 };
+
+export const guardianWeeklyCurrentSubscription: ProductDetail = {
+	mmaCategory: 'subscriptions',
+	tier: 'Guardian Weekly - Domestic',
+	isPaidTier: true,
+	selfServiceCancellation: {
+		isAllowed: false,
+		shouldDisplayEmail: false,
+		phoneRegionsToDisplay: ['UK & ROW'],
+	},
+	joinDate: '2021-12-15',
+	optIn: true,
+	subscription: {
+		paymentMethod: 'Card',
+		card: {
+			last4: '4242',
+			expiry: { month: 4, year: 2024 },
+			type: 'Visa',
+			stripePublicKeyForUpdate: 'pk_test_123',
+			email: 'test.user@example.com',
+		},
+		contactId: '0039E00001KA26BQAT',
+		deliveryAddress: {
+			addressLine1: 'Kings Place',
+			addressLine2: '90 York Way',
+			town: 'Canberra',
+			region: 'ACT',
+			postcode: '2601',
+			country: 'Australia',
+		},
+		safeToUpdatePaymentMethod: true,
+		start: '2021-12-24',
+		end: '2022-12-15',
+		nextPaymentPrice: 3250,
+		nextPaymentDate: '2021-12-24',
+		lastPaymentDate: null,
+		chargedThroughDate: null,
+		renewalDate: '2022-12-15',
+		anniversaryDate: '2022-12-24',
+		cancelledAt: false,
+		subscriberId: 'A-S00293857',
+		subscriptionId: 'A-S00293857',
+		trialLength: -37,
+		autoRenew: true,
+		plan: {
+			name: 'Guardian Weekly - Domestic',
+			amount: 3250,
+			currency: '$',
+			currencyISO: 'AUD',
+			interval: 'month',
+		},
+		currentPlans: [
+			{
+				name: null,
+				start: '2021-12-24',
+				end: '2022-12-15',
+				shouldBeVisible: true,
+				chargedThrough: null,
+				amount: 3250,
+				currency: '$',
+				currencyISO: 'AUD',
+				interval: 'month',
+			},
+		],
+		futurePlans: [],
+		readerType: 'Direct',
+		accountId: '8ad0965d7dbcc507017dbe20afd33ac4',
+		deliveryAddressChangeEffectiveDate: '2022-02-11',
+	},
+	isTestUser: false,
+};
+
+export const contribution: ProductDetail = {
+	mmaCategory: 'contributions',
+	tier: 'Contributor',
+	isPaidTier: true,
+	selfServiceCancellation: {
+		isAllowed: true,
+		shouldDisplayEmail: true,
+		phoneRegionsToDisplay: ['UK & ROW', 'US', 'AUS'],
+	},
+	joinDate: '2022-01-05',
+	optIn: true,
+	subscription: {
+		paymentMethod: 'PayPal',
+		payPalEmail: 'sb-ltpuy8454870@personal.example.com',
+		contactId: '0039E00001KA26BQAT',
+		deliveryAddress: {
+			addressLine1: 'Kings Place',
+			addressLine2: '90 York Place',
+			town: 'Canberra',
+			region: 'ACT',
+			postcode: '2601',
+			country: 'Australia',
+		},
+		safeToUpdatePaymentMethod: true,
+		start: '2022-01-05',
+		end: '2022-02-05',
+		nextPaymentPrice: 300,
+		nextPaymentDate: '2022-02-05',
+		lastPaymentDate: '2022-01-05',
+		chargedThroughDate: '2022-02-05',
+		renewalDate: '2023-01-05',
+		anniversaryDate: '2023-01-05',
+		cancelledAt: false,
+		subscriberId: 'A-S00303370',
+		subscriptionId: 'A-S00303370',
+		trialLength: -24,
+		autoRenew: true,
+		plan: {
+			name: 'Contributor',
+			amount: 300,
+			currency: '£',
+			currencyISO: 'GBP',
+			interval: 'month',
+		},
+		currentPlans: [
+			{
+				name: null,
+				start: '2022-01-05',
+				end: '2023-01-05',
+				shouldBeVisible: true,
+				chargedThrough: '2022-02-05',
+				amount: 300,
+				currency: '£',
+				currencyISO: 'GBP',
+				interval: 'month',
+			},
+		],
+		futurePlans: [],
+		readerType: 'Direct',
+		accountId: '8ad09f8a7e25bda3017e296317464818',
+	},
+	isTestUser: false,
+};
