--- app/cypress/fixtures/fixtures.ts
@@ -1,681 +1,0 @@
-import { StripeSetupIntent } from '../../shared/stripeSetupIntent';
-
-export const stripePaymentMethod = {
-	id: 'pm_1J2RZI2eZvKYlo2CaP803dMY',
-	object: 'payment_method',
-	billing_details: {
-		address: {
-			city: 'Kuala Lumpur',
-			country: null,
-			line1: 'Level 10, 1 Sentral, Jalan Rakyat, Kuala Lumpur Sentral',
-			line2: null,
-			postal_code: '50706',
-			state: null,
-		},
-		email: 'jenny@example.com',
-		name: null,
-		phone: '+15555555555',
-	},
-	card: {
-		brand: 'visa',
-		checks: {
-			address_line1_check: null,
-			address_postal_code_check: null,
-			cvc_check: 'pass',
-		},
-		country: 'US',
-		exp_month: 8,
-		exp_year: 2022,
-		fingerprint: 'Xt5EWLLDS7FJjR1c',
-		funding: 'credit',
-		generated_from: null,
-		last4: '4242',
-		networks: {
-			available: ['visa'],
-			preferred: null,
-		},
-		three_d_secure_usage: {
-			supported: true,
-		},
-		wallet: null,
-	},
-	created: 123456789,
-	customer: null,
-	livemode: false,
-	metadata: {
-		order_id: '123456789',
-	},
-	type: 'card',
-};
-
-export const ddPaymentMethod = {
-	accountName: 'JON R HEE',
-	accountNumber: '****9911',
-	sortCode: '200000',
-};
-
-export const stripeSetupIntent: StripeSetupIntent = {
-	id: 'seti_0KLFtUItVxyc3M6nBXnYb2jO',
-	client_secret:
-		'seti_0KLFtUItVxyc3Q6nBXnYb2jO_secret_L1IUioSNMNThetlMQnVtbCJu0Gj2cq1M',
-};
-
-export const confirmCardSetupResponse = {
-	status: 'succeeded',
-};
-
-export const executePaymentUpdateResponse = {
-	type: 'visa',
-	last4: '4242',
-	expiryMonth: 4,
-	expiryYear: 2024,
-};
-
-export const digitalDD = [
-	{
-		mmaCategory: 'subscriptions',
-		tier: 'Digital Pack',
-		isPaidTier: true,
-		selfServiceCancellation: {
-			isAllowed: false,
-			shouldDisplayEmail: false,
-			phoneRegionsToDisplay: ['UK & ROW'],
-		},
-		joinDate: '2021-11-11',
-		subscription: {
-			paymentMethod: 'DirectDebit',
-			mandate: {
-				accountName: 'Jon Hee',
-				accountNumber: '****9911',
-				sortCode: '200000',
-			},
-			contactId: '0039E00001KA26BQAT',
-			deliveryAddress: {
-				addressLine1: '11 Valnay Street',
-				addressLine2: '',
-				town: 'London',
-				postcode: 'SW17 8PS',
-				country: 'United Kingdom',
-			},
-			safeToUpdatePaymentMethod: true,
-			start: '2021-11-27',
-			end: '2022-11-11',
-			nextPaymentPrice: 599,
-			nextPaymentDate: '2021-11-27',
-			lastPaymentDate: null,
-			chargedThroughDate: null,
-			renewalDate: '2022-11-11',
-			anniversaryDate: '2022-11-27',
-			cancelledAt: false,
-			subscriptionId: 'A-S00278175',
-			trialLength: 12,
-			autoRenew: true,
-			currentPlans: [],
-			futurePlans: [
-				{
-					name: null,
-					start: '2021-11-27',
-					shouldBeVisible: true,
-				},
-			],
-			readerType: 'Direct',
-		},
-		isTestUser: false,
-	},
-];
-
-export const contribution = [
-	{
-		mmaCategory: 'contributions',
-		tier: 'Contributor',
-		isPaidTier: true,
-		selfServiceCancellation: {
-			isAllowed: true,
-			shouldDisplayEmail: true,
-			phoneRegionsToDisplay: ['UK & ROW', 'US', 'AUS'],
-		},
-		joinDate: '2022-01-05',
-		optIn: true,
-		subscription: {
-			paymentMethod: 'PayPal',
-			payPalEmail: 'sb-ltpuy8454870@personal.example.com',
-			contactId: '0039E00001KA26BQAT',
-			deliveryAddress: {
-				addressLine1: '11 Valnay Street',
-				addressLine2: '',
-				town: 'Canberra',
-				region: 'ACT',
-				postcode: '2601',
-				country: 'Australia',
-			},
-			safeToUpdatePaymentMethod: true,
-			start: '2022-01-05',
-			end: '2022-02-05',
-			nextPaymentPrice: 300,
-			nextPaymentDate: '2022-02-05',
-			lastPaymentDate: '2022-01-05',
-			chargedThroughDate: '2022-02-05',
-			renewalDate: '2023-01-05',
-			anniversaryDate: '2023-01-05',
-			cancelledAt: false,
-			subscriberId: 'A-S00303370',
-			subscriptionId: 'A-S00303370',
-			trialLength: -24,
-			autoRenew: true,
-			plan: {
-				name: 'Contributor',
-				amount: 300,
-				currency: '£',
-				currencyISO: 'GBP',
-				interval: 'month',
-			},
-			currentPlans: [
-				{
-					name: null,
-					start: '2022-01-05',
-					end: '2023-01-05',
-					shouldBeVisible: true,
-					chargedThrough: '2022-02-05',
-					amount: 300,
-					currency: '£',
-					currencyISO: 'GBP',
-					interval: 'month',
-				},
-			],
-			futurePlans: [],
-			readerType: 'Direct',
-			accountId: '8ad09f8a7e25bda3017e296317464818',
-			cancellationEffectiveDate: null,
-		},
-		isTestUser: false,
-	},
-];
-
-export const mmaResponse = [
-	{
-		mmaCategory: 'subscriptions',
-		tier: 'Digital Pack',
-		isPaidTier: true,
-		selfServiceCancellation: {
-			isAllowed: false,
-			shouldDisplayEmail: false,
-			phoneRegionsToDisplay: ['UK & ROW'],
-		},
-		joinDate: '2021-12-02',
-		optIn: true,
-		subscription: {
-			paymentMethod: 'DirectDebit',
-			account: { accountName: 'afsddfsa adfsdfsa' },
-			mandate: {
-				accountName: 'afsddfsa adfsdfsa',
-				accountNumber: '****9911',
-				sortCode: '200000',
-			},
-			contactId: '0039E00001KA26BQAT',
-			deliveryAddress: {
-				addressLine1: '11 Valnay Street',
-				addressLine2: '',
-				town: 'Canberra',
-				region: 'ACT',
-				postcode: '2601',
-				country: 'Australia',
-			},
-			safeToUpdatePaymentMethod: true,
-			start: '2021-12-18',
-			end: '2022-12-02',
-			nextPaymentPrice: 9900,
-			nextPaymentDate: '2021-12-18',
-			lastPaymentDate: null,
-			chargedThroughDate: null,
-			renewalDate: '2022-12-02',
-			anniversaryDate: '2022-12-18',
-			cancelledAt: false,
-			subscriberId: 'A-S00287957',
-			subscriptionId: 'A-S00287957',
-			trialLength: -42,
-			autoRenew: true,
-			plan: {
-				name: 'Digital Pack',
-				amount: 11900,
-				currency: '£',
-				currencyISO: 'GBP',
-				interval: 'year',
-			},
-			currentPlans: [
-				{
-					name: null,
-					start: '2021-12-18',
-					end: '2022-12-02',
-					shouldBeVisible: true,
-					chargedThrough: null,
-					amount: 11900,
-					currency: '£',
-					currencyISO: 'GBP',
-					interval: 'year',
-				},
-			],
-			futurePlans: [],
-			readerType: 'Direct',
-			accountId: '8ad08c0f7d768472017d7bc3e5960b20',
-			cancellationEffectiveDate: null,
-		},
-		isTestUser: false,
-	},
-	{
-		mmaCategory: 'subscriptions',
-		tier: 'Guardian Weekly - Domestic',
-		isPaidTier: true,
-		selfServiceCancellation: {
-			isAllowed: false,
-			shouldDisplayEmail: false,
-			phoneRegionsToDisplay: ['UK & ROW'],
-		},
-		joinDate: '2021-12-15',
-		optIn: true,
-		subscription: {
-			paymentMethod: 'Card',
-			card: {
-				last4: '4242',
-				expiry: { month: 4, year: 2024 },
-				type: 'Visa',
-				stripePublicKeyForUpdate: 'pk_test_Qm3CGRdrV4WfGYCpm0sftR0f',
-				email: 'jon.flynn+code@guardian.co.uk',
-			},
-			contactId: '0039E00001KA26BQAT',
-			deliveryAddress: {
-				addressLine1: '11 Valnay Street',
-				addressLine2: '',
-				town: 'Canberra',
-				region: 'ACT',
-				postcode: '2601',
-				country: 'Australia',
-			},
-			safeToUpdatePaymentMethod: true,
-			start: '2021-12-24',
-			end: '2022-12-15',
-			nextPaymentPrice: 3250,
-			nextPaymentDate: '2021-12-24',
-			lastPaymentDate: null,
-			chargedThroughDate: null,
-			renewalDate: '2022-12-15',
-			anniversaryDate: '2022-12-24',
-			cancelledAt: false,
-			subscriberId: 'A-S00293857',
-			subscriptionId: 'A-S00293857',
-			trialLength: -36,
-			autoRenew: true,
-			plan: {
-				name: 'Guardian Weekly - Domestic',
-				amount: 3250,
-				currency: '$',
-				currencyISO: 'AUD',
-				interval: 'month',
-			},
-			currentPlans: [
-				{
-					name: null,
-					start: '2021-12-24',
-					end: '2022-12-15',
-					shouldBeVisible: true,
-					chargedThrough: null,
-					amount: 3250,
-					currency: '$',
-					currencyISO: 'AUD',
-					interval: 'month',
-				},
-			],
-			futurePlans: [],
-			readerType: 'Direct',
-			accountId: '8ad0965d7dbcc507017dbe20afd33ac4',
-			cancellationEffectiveDate: null,
-			deliveryAddressChangeEffectiveDate: '2022-02-11',
-		},
-		isTestUser: false,
-	},
-	{
-		mmaCategory: 'contributions',
-		tier: 'Contributor',
-		isPaidTier: true,
-		selfServiceCancellation: {
-			isAllowed: true,
-			shouldDisplayEmail: true,
-			phoneRegionsToDisplay: ['UK & ROW', 'US', 'AUS'],
-		},
-		joinDate: '2022-01-29',
-		optIn: true,
-		subscription: {
-			paymentMethod: 'Card',
-			card: {
-				last4: '4242',
-				expiry: { month: 4, year: 2024 },
-				type: 'Visa',
-				stripePublicKeyForUpdate: 'pk_test_Qm3CGRdrV4WfGYCpm0sftR0f',
-				email: 'jon.flynn+code@guardian.co.uk',
-			},
-			contactId: '0039E00001KA26BQAT',
-			deliveryAddress: {
-				addressLine1: '11 Valnay Street',
-				addressLine2: '',
-				town: 'Canberra',
-				region: 'ACT',
-				postcode: '2601',
-				country: 'Australia',
-			},
-			safeToUpdatePaymentMethod: true,
-			start: '2022-01-29',
-			end: '2022-02-28',
-			nextPaymentPrice: 900,
-			nextPaymentDate: '2022-02-28',
-			lastPaymentDate: '2022-01-29',
-			chargedThroughDate: '2022-02-28',
-			renewalDate: '2023-01-29',
-			anniversaryDate: '2023-01-29',
-			cancelledAt: false,
-			subscriberId: 'A-S00314561',
-			subscriptionId: 'A-S00314561',
-			trialLength: 0,
-			autoRenew: true,
-			plan: {
-				name: 'Contributor',
-				amount: 900,
-				currency: '£',
-				currencyISO: 'GBP',
-				interval: 'month',
-			},
-			currentPlans: [
-				{
-					name: null,
-					start: '2022-01-29',
-					end: '2023-01-29',
-					shouldBeVisible: true,
-					chargedThrough: '2022-02-28',
-					amount: 900,
-					currency: '£',
-					currencyISO: 'GBP',
-					interval: 'month',
-				},
-			],
-			futurePlans: [],
-			readerType: 'Direct',
-			accountId: '8ad0965d7ea28ecf017ea765ec9522d5',
-			cancellationEffectiveDate: null,
-		},
-		isTestUser: false,
-	},
-	{
-		mmaCategory: 'subscriptions',
-		tier: 'Newspaper Digital Voucher',
-		isPaidTier: true,
-		selfServiceCancellation: {
-			isAllowed: false,
-			shouldDisplayEmail: false,
-			phoneRegionsToDisplay: ['UK & ROW'],
-		},
-		joinDate: '2021-12-15',
-		optIn: true,
-		subscription: {
-			paymentMethod: 'Card',
-			card: {
-				last4: '4242',
-				expiry: { month: 4, year: 2024 },
-				type: 'Visa',
-				stripePublicKeyForUpdate: 'pk_test_Qm3CGRdrV4WfGYCpm0sftR0f',
-				email: 'jon.flynn+code@guardian.co.uk',
-			},
-			contactId: '0039E00001KA26BQAT',
-			deliveryAddress: {
-				addressLine1: '11 Valnay Street',
-				addressLine2: '',
-				town: 'Canberra',
-				region: 'ACT',
-				postcode: '2601',
-				country: 'Australia',
-			},
-			safeToUpdatePaymentMethod: true,
-			start: '2021-12-23',
-			end: '2022-01-23',
-			nextPaymentPrice: 5299,
-			nextPaymentDate: '2022-01-23',
-			lastPaymentDate: '2021-12-23',
-			chargedThroughDate: '2022-01-23',
-			renewalDate: '2022-12-15',
-			anniversaryDate: '2022-12-23',
-			cancelledAt: false,
-			subscriberId: 'A-S00293815',
-			subscriptionId: 'A-S00293815',
-			trialLength: -37,
-			autoRenew: true,
-			plan: {
-				name: 'Newspaper Digital Voucher',
-				amount: 5299,
-				currency: '£',
-				currencyISO: 'GBP',
-				interval: 'month',
-			},
-			currentPlans: [
-				{
-					name: 'Everyday',
-					start: '2021-12-23',
-					end: '2022-12-15',
-					shouldBeVisible: true,
-					chargedThrough: '2022-01-23',
-					amount: 5299,
-					currency: '£',
-					currencyISO: 'GBP',
-					interval: 'month',
-					daysOfWeek: [
-						'Monday',
-						'Tuesday',
-						'Wednesday',
-						'Thursday',
-						'Friday',
-						'Saturday',
-						'Sunday',
-					],
-				},
-			],
-			futurePlans: [],
-			readerType: 'Direct',
-			accountId: '8ad09f8a7dbcc50c017dbda3bdc05f8d',
-			cancellationEffectiveDate: null,
-		},
-		isTestUser: false,
-	},
-];
-
-export const cancelledResponse = [
-	{
-		mmaCategory: 'contributions',
-		tier: 'Contributor',
-		subscription: {
-			subscriptionId: 'A-S00268736',
-			cancellationEffectiveDate: '2021-11-21',
-			start: '2021-10-21',
-			end: '2021-11-21',
-			readerType: 'Direct',
-			accountId: '8ad084a67ca1d07c017ca463884c5c46',
-		},
-	},
-	{
-		mmaCategory: 'contributions',
-		tier: 'Contributor',
-		subscription: {
-			subscriptionId: 'A-S00277393',
-			cancellationEffectiveDate: '2021-12-09',
-			start: '2021-11-09',
-			end: '2021-12-09',
-			readerType: 'Direct',
-			accountId: '8ad084a67cfec5bc017d0573d8da4641',
-		},
-	},
-	{
-		mmaCategory: 'contributions',
-		tier: 'Contributor',
-		subscription: {
-			subscriptionId: 'A-S00271109',
-			cancellationEffectiveDate: '2021-11-26',
-			start: '2021-10-26',
-			end: '2021-11-26',
-			readerType: 'Direct',
-			accountId: '8ad08c0f7cbcc1b1017cbed982091a59',
-		},
-	},
-	{
-		mmaCategory: 'contributions',
-		tier: 'Contributor',
-		subscription: {
-			subscriptionId: 'A-S00274043',
-			cancellationEffectiveDate: '2021-12-02',
-			start: '2021-11-02',
-			end: '2021-12-02',
-			readerType: 'Direct',
-			accountId: '8ad08c0f7cda94a9017ce0028fcb42d9',
-		},
-	},
-	{
-		mmaCategory: 'contributions',
-		tier: 'Contributor',
-		subscription: {
-			subscriptionId: 'A-S00274524',
-			cancellationEffectiveDate: '2021-12-03',
-			start: '2021-11-03',
-			end: '2021-12-03',
-			readerType: 'Direct',
-			accountId: '8ad0965d7ce3ed3b017ce563a4e10f13',
-		},
-	},
-	{
-		mmaCategory: 'contributions',
-		tier: 'Contributor',
-		subscription: {
-			subscriptionId: 'A-S00281897',
-			cancellationEffectiveDate: '2021-12-19',
-			start: '2021-11-19',
-			end: '2021-12-19',
-			readerType: 'Direct',
-			accountId: '8ad0965d7d323130017d37e7754379f4',
-		},
-	},
-	{
-		mmaCategory: 'contributions',
-		tier: 'Contributor',
-		subscription: {
-			subscriptionId: 'A-S00273719',
-			cancellationEffectiveDate: '2021-12-01',
-			start: '2021-11-01',
-			end: '2021-12-01',
-			readerType: 'Direct',
-			accountId: '8ad098317cdaa22b017cdc6dd4d846fb',
-		},
-	},
-	{
-		mmaCategory: 'contributions',
-		tier: 'Contributor',
-		subscription: {
-			subscriptionId: 'A-S00274063',
-			cancellationEffectiveDate: '2021-12-02',
-			start: '2021-11-02',
-			end: '2021-12-02',
-			readerType: 'Direct',
-			accountId: '8ad098317cdaa22b017ce01d1c634e11',
-		},
-	},
-	{
-		mmaCategory: 'contributions',
-		tier: 'Contributor',
-		subscription: {
-			subscriptionId: 'A-S00293413',
-			cancellationEffectiveDate: '2022-01-14',
-			start: '2021-12-14',
-			end: '2022-01-14',
-			readerType: 'Direct',
-			accountId: '8ad098317db2e76b017db915c0677859',
-		},
-	},
-	{
-		mmaCategory: 'contributions',
-		tier: 'Contributor',
-		subscription: {
-			subscriptionId: 'A-S00294335',
-			cancellationEffectiveDate: '2022-01-16',
-			start: '2021-12-16',
-			end: '2022-01-16',
-			readerType: 'Direct',
-			accountId: '8ad098317dc2ae8f017dc37bc2c63f94',
-		},
-	},
-];
-
-export const gwProductDetail = [
-	{
-		mmaCategory: 'subscriptions',
-		tier: 'Guardian Weekly - Domestic',
-		isPaidTier: true,
-		selfServiceCancellation: {
-			isAllowed: false,
-			shouldDisplayEmail: false,
-			phoneRegionsToDisplay: ['UK & ROW'],
-		},
-		joinDate: '2021-12-15',
-		optIn: true,
-		subscription: {
-			paymentMethod: 'Card',
-			card: {
-				last4: '4242',
-				expiry: { month: 4, year: 2024 },
-				type: 'Visa',
-				stripePublicKeyForUpdate: 'pk_test_Qm3CGRdrV4WfGYCpm0sftR0f',
-				email: 'jon.flynn+code@guardian.co.uk',
-			},
-			contactId: '0039E00001KA26BQAT',
-			deliveryAddress: {
-				addressLine1: '11 Valnay Street',
-				addressLine2: '',
-				town: 'Canberra',
-				region: 'ACT',
-				postcode: '2601',
-				country: 'Australia',
-			},
-			safeToUpdatePaymentMethod: true,
-			start: '2021-12-24',
-			end: '2022-12-15',
-			nextPaymentPrice: 3250,
-			nextPaymentDate: '2021-12-24',
-			lastPaymentDate: null,
-			chargedThroughDate: null,
-			renewalDate: '2022-12-15',
-			anniversaryDate: '2022-12-24',
-			cancelledAt: false,
-			subscriberId: 'A-S00293857',
-			subscriptionId: 'A-S00293857',
-			trialLength: -37,
-			autoRenew: true,
-			plan: {
-				name: 'Guardian Weekly - Domestic',
-				amount: 3250,
-				currency: '$',
-				currencyISO: 'AUD',
-				interval: 'month',
-			},
-			currentPlans: [
-				{
-					name: null,
-					start: '2021-12-24',
-					end: '2022-12-15',
-					shouldBeVisible: true,
-					chargedThrough: null,
-					amount: 3250,
-					currency: '$',
-					currencyISO: 'AUD',
-					interval: 'month',
-				},
-			],
-			futurePlans: [],
-			readerType: 'Direct',
-			accountId: '8ad0965d7dbcc507017dbe20afd33ac4',
-			cancellationEffectiveDate: null,
-			deliveryAddressChangeEffectiveDate: '2022-02-11',
-		},
-		isTestUser: false,
-	},
-];
