+++ app/cypress/integration/parallel-1/updatePaymentDetails.spec.ts
@@ -1,6 +1,14 @@
 import { setLocalBaseUrl } from '../../lib/setLocalBaseUrl';
-import * as fixtures from '../../fixtures/fixtures';
-// import { PaymentMethod } from "../../../client/components/payment/update/updatePaymentFlow";
+import {
+	guardianWeeklyCurrentSubscription,
+	digitalDD,
+} from '../../../client/fixtures/productDetail';
+import {
+	stripeSetupIntent,
+	executePaymentUpdateResponse,
+	ddPaymentMethod,
+} from '../../../client/fixtures/payment';
+import { paymentMethods } from '../../../client/fixtures/stripe';
 
 // 'No IAB consent management framework' exception is thrown from here: https://github.com/guardian/consent-management-platform/blob/405a4fee4c54c2bdabea3df0fd1bf187ae6d7927/src/onConsentChange.ts#L34
 Cypress.on('uncaught:exception', () => {
@@ -17,12 +25,12 @@ describe('E2E Page rendering', function () {
 
 			cy.intercept('GET', '/api/me/mma', {
 				statusCode: 200,
-				body: fixtures.mmaResponse,
+				body: [guardianWeeklyCurrentSubscription],
 			}).as('mma');
 
 			cy.intercept('GET', '/api/cancelled/', {
 				statusCode: 200,
-				body: fixtures.cancelledResponse,
+				body: [],
 			}).as('cancelled');
 
 			cy.wait(1000);
@@ -46,32 +54,32 @@ describe('E2E Page rendering', function () {
 	it('Complete card payment update', function () {
 		cy.intercept('GET', '/api/me/mma?productType=*', {
 			statusCode: 200,
-			body: fixtures.gwProductDetail,
+			body: [guardianWeeklyCurrentSubscription],
 		}).as('product_detail');
 
 		cy.intercept('GET', '/api/me/mma/**', {
 			statusCode: 200,
-			body: fixtures.gwProductDetail,
+			body: [guardianWeeklyCurrentSubscription],
 		}).as('refetch_subscription');
 
 		cy.intercept('POST', '/api/payment/card', {
 			statusCode: 200,
-			body: fixtures.stripeSetupIntent,
+			body: stripeSetupIntent,
 		}).as('createSetupIntent');
 
 		cy.intercept('POST', '/api/payment/card/**', {
 			statusCode: 200,
-			body: fixtures.executePaymentUpdateResponse,
+			body: executePaymentUpdateResponse,
 		}).as('scala_backend');
 
 		cy.intercept('POST', 'https://api.stripe.com/v1/setup_intents/**', {
 			statusCode: 200,
-			body: fixtures.confirmCardSetupResponse,
+			body: { status: 'succeeded' },
 		}).as('confirmCardSetup');
 
 		cy.intercept('POST', 'https://api.stripe.com/v1/payment_methods', {
 			statusCode: 200,
-			body: fixtures.stripePaymentMethod,
+			body: paymentMethods,
 		});
 
 		cy.visit('/payment/subscriptioncard');
@@ -110,92 +118,10 @@ describe('E2E Page rendering', function () {
 		cy.get('@scala_backend.all').should('have.length', 1);
 	});
 
-	it('Completes adding holiday stop', function () {
-		cy.intercept('GET', '/api/me/mma', {
-			statusCode: 200,
-			body: fixtures.gwProductDetail,
-		}).as('mma');
-
-		cy.intercept('GET', '/api/me/mma?productType=Weekly', {
-			statusCode: 200,
-			body: fixtures.gwProductDetail,
-		}).as('product_detail');
-
-		cy.intercept('GET', '/api/holidays/*/*', {
-			statusCode: 200,
-			body: {
-				nextInvoiceDateAfterToday: '2022-02-24',
-				potentials: [
-					{
-						publicationDate: '2022-03-25',
-						credit: -8.13,
-						invoiceDate: '2022-04-24',
-					},
-				],
-			},
-		}).as('fetch_holidays');
-
-		cy.intercept('GET', '/api/holidays/*', {
-			statusCode: 200,
-			body: {
-				existing: [
-					{
-						id: 'a2k9E000005NnbrQAC',
-						startDate: '2022-03-11',
-						endDate: '2022-03-12',
-						subscriptionName: 'A-S00293857',
-						publicationsImpacted: [
-							{
-								publicationDate: '2022-03-11',
-								estimatedPrice: -8.13,
-								invoiceDate: '2022-03-24',
-								isActioned: false,
-							},
-						],
-						mutabilityFlags: {
-							isFullyMutable: true,
-							isEndDateEditable: true,
-						},
-					},
-				],
-				issueSpecifics: [
-					{
-						firstAvailableDate: '2022-02-05',
-						issueDayOfWeek: 5,
-					},
-				],
-				annualIssueLimit: 6,
-				firstAvailableDate: '2022-02-05',
-			},
-		}).as('fetch_holidays');
-
-		cy.intercept('POST', '/api/holidays', {
-			statusCode: 200,
-			body: {
-				message: 'success',
-			},
-		}).as('create_holiday_stop');
-
-		cy.visit('/suspend/guardianweekly');
-		cy.wait('@fetch_holidays');
-		cy.wait('@product_detail');
-		cy.get('[data-cy="create_suspension_cta"] button').click();
-
-		cy.findByText('Choose the dates you will be away');
-		cy.get('[data-cy="date_picker"] div').eq(8).click();
-		cy.get('[data-cy="date_picker"] div').eq(10).click();
-		cy.findByText('Review details').click();
-
-		cy.findByText('Confirm').click();
-
-		cy.wait('@create_holiday_stop');
-		cy.findByText('Your schedule has been set');
-	});
-
 	it('Shows correct error messages for direct debit form', function () {
 		cy.intercept('GET', '/api/me/mma?productType=*', {
 			statusCode: 200,
-			body: fixtures.digitalDD,
+			body: [digitalDD],
 		}).as('product_detail');
 
 		cy.visit('/payment/digital');
@@ -231,17 +157,17 @@ describe('E2E Page rendering', function () {
 	it('Complete direct debit payment update', function () {
 		cy.intercept('GET', '/api/me/mma?productType=*', {
 			statusCode: 200,
-			body: fixtures.digitalDD,
+			body: [digitalDD],
 		}).as('product_detail');
 
 		cy.intercept('GET', '/api/me/mma/**', {
 			statusCode: 200,
-			body: fixtures.digitalDD,
+			body: [digitalDD],
 		}).as('refetch_subscription');
 
 		cy.intercept('POST', '/api/payment/dd/**', {
 			statusCode: 200,
-			body: fixtures.ddPaymentMethod,
+			body: ddPaymentMethod,
 		}).as('scala_backend');
 
 		cy.visit('/payment/digital');
@@ -271,7 +197,7 @@ describe('E2E Page rendering', function () {
 	it('Show recaptcha error', function () {
 		cy.intercept('GET', '/api/me/mma?productType=*', {
 			statusCode: 200,
-			body: fixtures.gwProductDetail,
+			body: [guardianWeeklyCurrentSubscription],
 		}).as('product_detail');
 
 		cy.visit('/payment/subscriptioncard');
